LayoutTests/fast/forms/date/date-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for date input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
PASS stepUp("2010-02-10", null, null) is "2010-02-11"
PASS stepDown("2010-02-10", null, null) is "2010-02-09"
PASS stepUp("2010-02-10", null, null, 10) is "2010-02-20"
PASS stepDown("2010-02-10", null, null, 11) is "2010-01-30"
PASS stepUp("1970-01-01", "4", null, 2) is "1970-01-09"
PASS stepDown("1970-01-01", "4", null, 3) is "1969-12-20"
Step=any
PASS stepUp("2010-02-10", "any", null) is "2010-02-11"
PASS stepDown("2010-02-10", "any", null) is "2010-02-09"
Overflow/underflow
PASS stepUp("2010-02-10", "3.40282346e+38", null) is "275760-09-13"
PASS stepDown("2010-02-10", "3.40282346e+38", null) is "1970-01-01"
PASS stepUp("2010-02-10", "1", "2010-02-10") is "2010-02-10"
PASS stepDown("2010-02-10", "1", "2010-02-10") is "2010-02-10"
stepDown()/stepUp() for stepMismatch values
PASS stepDown("2010-02-10", "3", "2010-02-06") is "2010-02-09"
PASS stepUp("1970-01-02", "2", "") is "1970-01-03"

PASS successfullyParsed is true

TEST COMPLETE
